# Intrinsics optimizations.