--- a/src/NewId.Tests/NewId.Tests.csproj
+++ b/src/NewId.Tests/NewId.Tests.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFrameworks>netcoreapp3.1;net6.0</TargetFrameworks>
     <RootNamespace>MassTransit.NewIdTests</RootNamespace>
   </PropertyGroup>