--- a/src/NewId/NewId.csproj
+++ b/src/NewId/NewId.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
+    <TargetFrameworks>netstandard2.0;net6.0</TargetFrameworks>
     <RootNamespace>MassTransit</RootNamespace>
   </PropertyGroup>