@phatboyg 2023-01-02
Nice work, I'll merge this once I convert the repository to build on GitHub Actions. I haven't used AppVeyor in over a year, so I'm not sure it even still works.

I'm also copying these changes into MassTransit.Abstractions (unless you want to submit a similar PR to that repository so that you get credit in the commit history for your work.