@TimothyMakkison 2023-01-03
I think I've found a bug with `FromSequentialGuid` I think it incorrectly swaps the bytes of the tick counter. Tests won't detect this because the first NewId generated is always used with the sequence number 0x0000. I'm not 100% sure if this is intended or even a bug.

Note that this possible bug existed before my changes although I did transfer this behaviour into intrinsics.

### Example failing test
```C#
[Test]
public void Should_parse_sequential_guid_2_as_newid()
{
    NewId n = NewId.Next(2)[1];

    var nn = n.ToGuid();
    var g = n.ToSequentialGuid();

    var ng = NewId.FromSequentialGuid(g);

    Assert.AreEqual(n, ng);

    // Also checks to see if this would throw
    Assert.IsTrue(ng.Timestamp != default);
}
```

### Seemingly fixed with 
```C#
static void FromSequentialByteArray(in byte[] bytes, out Int32 a, out Int32 b, out Int32 c, out Int32 d)
{
    a = bytes[3] << 24 | bytes[2] << 16 | bytes[1] << 8 | bytes[0];
    b = bytes[5] << 24 | bytes[4] << 16 | bytes[7] << 8 | bytes[6];
    c = bytes[8] << 24 | bytes[9] << 16 | bytes[10] << 8 | bytes[11];
    d = bytes[12] << 24 | bytes[13] << 16 | bytes[14] << 8 | bytes[15];
}
```